requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
librosa==0.10.1
essentia==2.1b6.dev1034
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
pydub==0.25.1
soundfile==0.12.1
matplotlib==3.7.2
seaborn==0.12.2
python-dotenv==1.0.0
aiofiles==23.2.1
